+++ app/src/main/java/com/hindou/smash/Controllers/CancelReservationActivity.java
@@ -1,4 +1,4 @@
-package com.hindou.smash;
+package com.hindou.smash.Controllers;
 
 import android.content.Context;
 import android.content.Intent;
@@ -16,11 +16,10 @@ import com.android.volley.Request;
 import com.android.volley.Response;
 import com.android.volley.VolleyError;
 import com.android.volley.toolbox.StringRequest;
-import com.hindou.smash.Models.Bike;
 import com.hindou.smash.Models.Reservation;
 import com.hindou.smash.Models.User;
-import com.hindou.smash.adapter.BikeAdapter;
-import com.hindou.smash.adapter.ReservationAdapter;
+import com.hindou.smash.R;
+import com.hindou.smash.Adapters.CancelReservationAdapter;
 import com.hindou.smash.utils.GlobalVars;
 import com.hindou.smash.utils.SessionsManager;
 import com.hindou.smash.utils.VolleySingleton;
@@ -35,7 +34,7 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-public class LockBikesActivity extends AppCompatActivity {
+public class CancelReservationActivity extends AppCompatActivity {
 
     private SessionsManager sessionsManager;
     private User connectedUser;
@@ -43,14 +42,14 @@ public class LockBikesActivity extends AppCompatActivity {
     private Context mContext;
 
     private RecyclerView mReservationList;
-    private ReservationAdapter mAdapter;
+    private CancelReservationAdapter mAdapter;
     private List<Reservation> mReservationSrcList;
     private MaterialDialog mLoader;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_lock_bikes);
+        setContentView(R.layout.activity_cancel_reservation);
         sessionsManager = SessionsManager.getInstance(this);
 
         if (!sessionsManager.isActive()){
@@ -61,11 +60,12 @@ public class LockBikesActivity extends AppCompatActivity {
         }
     }
 
+
     private void initializeComponents(){
 
 
         //Set screen title
-        setTitle(R.string.lock_activity_title);
+        setTitle(R.string.cancel_activity_title);
 
         mContext = this;
         mReservationSrcList = new ArrayList<>();
@@ -80,9 +80,9 @@ public class LockBikesActivity extends AppCompatActivity {
             getSupportActionBar().setDisplayHomeAsUpEnabled(true);
         }
 
-        mReservationList = findViewById(R.id.reservation_list);
+        mReservationList = findViewById(R.id.reservation_list_cancel);
         mReservationList.setLayoutManager(new LinearLayoutManager(mContext, LinearLayoutManager.VERTICAL, false));
-        mAdapter = new ReservationAdapter(mReservationSrcList, mContext);
+        mAdapter = new CancelReservationAdapter(mReservationSrcList, mContext);
         mReservationList.setAdapter(mAdapter);
 
         mLoader = new MaterialDialog.Builder(this)
@@ -93,6 +93,8 @@ public class LockBikesActivity extends AppCompatActivity {
 
 
     private void getReservations(){
+
+
         StringRequest request = new StringRequest(Request.Method.POST, GlobalVars.API_URL + "getReservation.php",
                 new Response.Listener<String>() {
                     @Override
@@ -112,8 +114,10 @@ public class LockBikesActivity extends AppCompatActivity {
                                                     data.getJSONObject(i).getString("name"),
                                                     data.getJSONObject(i).getString("duration"),
                                                     data.getJSONObject(i).getString("timestamp")
+
                                             )
                                     );
+
 
                                 }
 
@@ -156,7 +160,7 @@ public class LockBikesActivity extends AppCompatActivity {
     }
 
     private void changeActivity(Class<?> destination, boolean flag) {
-        Intent intent = new Intent(LockBikesActivity.this, destination);
+        Intent intent = new Intent(CancelReservationActivity.this, destination);
         startActivity(intent);
         if(flag) finish();
     }
