+++ app/src/main/java/com/hindou/smash/Controllers/Ehealth.java
@@ -1,41 +1,26 @@
-package com.hindou.smash;
+package com.hindou.smash.Controllers;
 
-import android.bluetooth.BluetoothAdapter;
-import android.bluetooth.BluetoothDevice;
-import android.bluetooth.BluetoothSocket;
-import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
-import android.os.Handler;
 import android.support.design.widget.TabLayout;
 import android.support.v4.view.ViewPager;
 import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.Toolbar;
-import android.util.Log;
 import android.view.MenuItem;
 import android.view.View;
 import android.widget.Button;
-import android.widget.TextView;
-import android.widget.Toast;
 
-import com.afollestad.materialdialogs.MaterialDialog;
 import com.hindou.smash.Models.HealthInfo;
-import com.hindou.smash.Models.User;
 import com.hindou.smash.Models.WaterDrink;
-import com.hindou.smash.adapter.ReservationAdapter;
-import com.hindou.smash.adapter.ViewPagerAdapter;
-import com.hindou.smash.utils.SessionsManager;
-
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.OutputStream;
-import java.util.ArrayList;
-import java.util.UUID;
+import com.hindou.smash.R;
+import com.hindou.smash.Adapters.ViewPagerAdapter;
+import com.hindou.smash.Services.HealthServices;
 
 import io.realm.Realm;
-import io.realm.RealmConfiguration;
 import io.realm.RealmResults;
+
+import static com.google.android.gms.cast.CastRemoteDisplayLocalService.startService;
+import static com.google.android.gms.cast.CastRemoteDisplayLocalService.stopService;
 
 /**
  * Created by HP on 23/03/2018.
@@ -47,6 +32,7 @@ public class Ehealth extends AppCompatActivity {
     ViewPager viewPager;
     public RealmResults<HealthInfo> list;
     public RealmResults<WaterDrink> list2;
+    Button start, stop;
 
     private Realm realm;
 
@@ -58,7 +44,7 @@ public class Ehealth extends AppCompatActivity {
         //Set screen title
         setTitle(R.string.health_frag_title);
         //Init realm database
-        Realm.init(this);
+       Realm.init(this);
         Realm.getDefaultInstance();
         realm=Realm.getDefaultInstance();
         list = realm.where(HealthInfo.class).findAll();
@@ -76,6 +62,18 @@ public class Ehealth extends AppCompatActivity {
         tabLayout = (TabLayout) findViewById(R.id.tab_layout);
         tabLayout.setupWithViewPager(viewPager);
 
+
+        HealthServices.startstop=true;
+
+
+      /*  stop.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                stopService(new Intent(getBaseContext(),HealthServices.class));
+                HealthServices.startstop = false;
+            }
+        });
+*/
         viewPager.addOnPageChangeListener(new ViewPager.OnPageChangeListener() {
             @Override
             public void onPageScrolled(int position, float positionOffset, int positionOffsetPixels) {
