+++ app/src/main/java/com/hindou/smash/Controllers/ReserveActivity.java
@@ -1,4 +1,4 @@
-package com.hindou.smash;
+package com.hindou.smash.Controllers;
 
 import android.content.Context;
 import android.content.Intent;
@@ -18,7 +18,8 @@ import com.android.volley.VolleyError;
 import com.android.volley.toolbox.StringRequest;
 import com.hindou.smash.Models.Bike;
 import com.hindou.smash.Models.User;
-import com.hindou.smash.adapter.BikeAdapter;
+import com.hindou.smash.R;
+import com.hindou.smash.Adapters.BikeAdapter;
 import com.hindou.smash.utils.GlobalVars;
 import com.hindou.smash.utils.SessionsManager;
 import com.hindou.smash.utils.VolleySingleton;
