+++ app/src/main/java/com/hindou/smash/Controllers/StationRoutingActivity.java
@@ -1,4 +1,4 @@
-package com.hindou.smash;
+package com.hindou.smash.Controllers;
 
 import android.content.Intent;
 import android.support.v7.app.AppCompatActivity;
@@ -24,6 +24,7 @@ import com.google.android.gms.maps.model.MarkerOptions;
 import com.google.android.gms.maps.model.Polyline;
 import com.google.android.gms.maps.model.PolylineOptions;
 import com.hindou.smash.Models.User;
+import com.hindou.smash.R;
 
 
 import java.util.ArrayList;
