--- app/src/main/java/com/hindou/smash/adapter/ViewPagerAdapter.java
@@ -1,39 +1,0 @@
-package com.hindou.smash.adapter;
-
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentManager;
-import android.support.v4.app.FragmentPagerAdapter;
-
-import java.util.ArrayList;
-import java.util.List;
-
-/**
- * Created by HP on 12/04/2018.
- */
-
-public class ViewPagerAdapter extends FragmentPagerAdapter {
-    private final List<Fragment> fragmentList = new ArrayList<>();
-    private final List<String> fragmentTitleList = new ArrayList<>();
-
-    public ViewPagerAdapter(FragmentManager fm) {
-        super(fm);
-    }
-
-    @Override
-    public Fragment getItem(int position) {
-        return fragmentList.get(position);
-    }
-
-    @Override
-    public int getCount() {
-        return fragmentTitleList.size();
-    }
-    @Override
-    public CharSequence getPageTitle(int position) {
-        return fragmentTitleList.get(position);
-    }
-    public void addFragment(Fragment fragment, String title){
-        fragmentList.add(fragment);
-        fragmentTitleList.add(title);
-    }
-}
