+++ app/src/main/res/layout/activity_cancel_reservation.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.hindou.smash.CancelReservationActivity">
+    tools:context="com.hindou.smash.Controllers.CancelReservationActivity">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
