+++ app/src/main/res/layout/activity_event.xml
@@ -5,7 +5,7 @@ android:layout_width="match_parent"
 android:layout_height="match_parent"
 xmlns:tools="http://schemas.android.com/tools"
 xmlns:app="http://schemas.android.com/apk/res-auto"
-tools:context=".EventActivity"
+tools:context=".Controllers.EventActivity"
 android:background="@color/ButtonsTextColors">
 <android.support.design.widget.AppBarLayout
     android:layout_width="match_parent"
