+++ app/src/main/res/layout/activity_health_history.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     xmlns:tools="http://schemas.android.com/tools"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".HealthHistoryActivity"
+    tools:context=".Controllers.HealthHistoryActivity"
     android:background="@color/ButtonsTextColors">
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
