+++ app/src/main/res/layout/activity_lock_bikes.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.hindou.smash.LockBikesActivity">
+    tools:context="com.hindou.smash.Controllers.LockBikesActivity">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
