+++ app/src/main/res/layout/activity_login.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:custom="http://schemas.android.com/apk/res-auto"
-    tools:context="com.hindou.smash.LoginActivity"
+    tools:context="com.hindou.smash.Controllers.LoginActivity"
     android:id="@+id/loginholder"
     android:background="@color/ButtonsTextColors">
