+++ app/src/main/res/layout/activity_reserve.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context="com.hindou.smash.ReserveActivity">
+    tools:context="com.hindou.smash.Controllers.ReserveActivity">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
