+++ app/src/main/res/layout/activity_signup.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.hindou.smash.SignupActivity"
+    tools:context="com.hindou.smash.Controllers.SignupActivity"
     android:background="@color/ButtonsTextColors">
 
     <TextView
