+++ app/src/main/res/layout/activity_splash.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.hindou.smash.SplashActivity"
+    tools:context="com.hindou.smash.Controllers.SplashActivity"
   android:background="@color/tint">
 
     <ImageView
